Fix Java 8 compatibility issues. (#69)
Compiling the library with a JDK version >8 and trying to run it against a JVM
version <=8 creates compatibility issues. In this particular case, Java 9 has
introduced overridden methods with covariant return types for the following
methods in `java.nio.ByteBuffer` that are used by the library:

- position
- limit
- flip
- clear

In Java 9 they all now return `ByteBuffer`, whereas the methods they override
return `Buffer`, resulting in exceptions like this when executing on Java 8 and
lower:

java.lang.NoSuchMethodError: java.nio.ByteBuffer.limit(I)Ljava/nio/ByteBuffer

This is because the generated byte code includes the static return type of the
method, which is not found on Java 8 and lower because the overloaded methods
with covariant return types don't exist.

The solution is to cast `ByteBuffer` instances to `Buffer` before calling the
method.